fixed linting errors
