+++ resources/assets/components/utilities/CtaPopover/CtaPopoverButton.js
@@ -2,9 +2,23 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Button from './Button/Button.js';
 
-const button = ({ handleClick, buttonText, link }) => {
-  const handleClick = () => {};
+import { trackAnalyticsEvent } from '../../../helpers/analytics';
 
+const button = ({ handleClick, buttonText, link }) => {
+  const handleClick = () =>
+    trackAnalyticsEvent({
+      metadata: {
+        category: 'site_action',
+        target: 'button',
+        verb: 'clicked',
+        noun: 'call_to_action',
+        adjective: 'popover',
+        label: 'call_to_action_popover',
+      },
+      context: {
+        url: link,
+      },
+    });
   return;
 };
