+++ resources/assets/components/utilities/CtaPopover/cta-popover-button.scss
@@ -1,5 +1,3 @@
-@import '../../../scss/next-toolbox';
-
 .flexContainer {
   display: flex;
   border: none;
